pyproject.toml
[project]
name = "protocol-query"
version = "0.1.0"
description = "Interactive clinical protocol querying CLI with hybrid search and LLM-powered analysis"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Protocol Query Team" }]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pypdf>=4.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "sentence-transformers>=2.2.0",
    "sqlite-vec>=0.1.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "tqdm>=4.65.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
protocol-query = "protocol_query.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/protocol_query"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
